backend/requirements.txt
Flask==2.3.2
transformers==4.31.0
torch==2.0.1
Pillow==10.0.0
requests==2.31.0